chore: move Page to layout folder
